Don't put result of 'neg_GF()' back into the same variable. See issue #2.
Instead of using the same variable, use some other temporary variable. Care was
taken to choose such temporary variable that writing into it doesn't affect the
rest of the function (either variable is not used after that or it gets some new
value written into it before the next read).